// data/github/current/ui/packages/item-picker2/repository-assignable-users/__generated__/ItemPickerRepositoryAssignableUsersList_Query.graphql.ts
export type ItemPickerRepositoryAssignableUsersList_Query$variables = {
  count: number;
  owner: string;
  query?: string | null | undefined;
  repo: string;
};
export type ItemPickerRepositoryAssignableUsersList_Query$data = {
  readonly repository: {
    readonly " $fragmentSpreads": FragmentRefs<"ItemPickerRepositoryAssignableUsersList_Fragment">;
  } | null | undefined;
};
export type ItemPickerRepositoryAssignableUsersList_Query = {
  response: ItemPickerRepositoryAssignableUsersList_Query$data;
  variables: ItemPickerRepositoryAssignableUsersList_Query$variables;
};